Rendering Contexts

+++ src/main/kotlin/uk/co/jcox/chemvis/cvengine/RenderTarget.kt
@@ -1,9 +1,10 @@
 package uk.co.jcox.chemvis.cvengine
 
+
 data class RenderTarget(
     val frameBuffer: Int,
     val colourAttachmentTexture: Int,
     val depthAttachmentRenderBuffer: Int,
-    val width: Int,
-    val height: Int,
+    val width: Float,
+    val height: Float,
 )
